Refactor: Remove search typing delay
