+++ frontend/lib/js/concept-trees/ConceptTreeSearchBox.js
@@ -1,21 +1,67 @@
 // @flow
 
-import React from "react";
+import * as React from "react";
 import styled from "@emotion/styled";
 import T from "i18n-react";
 import { connect } from "react-redux";
 
-import { SearchBox } from "../form-components";
+import TransparentButton from "../button/TransparentButton";
+import IconButton from "../button/IconButton";
+
+import { isEmpty } from "../common/helpers";
+import AnimatedDots from "../common/components/AnimatedDots";
+
+import BaseInput from "../form-components/BaseInput";
+
+import type { DatasetIdT } from "../api/types";
 
 import { getAreTreesAvailable } from "./selectors";
-import {
-  searchTrees,
-  changeSearchQuery,
-  clearSearchQuery,
-  toggleShowMismatches
-} from "./actions";
+import { searchTrees, clearSearchQuery, toggleShowMismatches } from "./actions";
 
-import TransparentButton from "../button/TransparentButton";
+import type { SearchT } from "./reducer";
+
+const Root = styled("div")`
+  margin: 0 10px 5px;
+  position: relative;
+`;
+
+const StyledBaseInput = styled(BaseInput)`
+  width: 100%;
+  input {
+    width: 100%;
+    &::placeholder {
+      color: ${({ theme }) => theme.col.grayMediumLight};
+      opacity: 1;
+    }
+  }
+`;
+
+const Right = styled("div")`
+  position: absolute;
+  top: 0px;
+  right: 30px;
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  height: 36px;
+`;
+
+const StyledIconButton = styled(IconButton)`
+  color: ${({ theme }) => theme.col.gray};
+`;
+
+const TinyText = styled("p")`
+  margin: 3px 0;
+  font-size: ${({ theme }) => theme.font.xs};
+  color: ${({ theme }) => theme.col.gray};
+`;
+
+const Row = styled("div")`
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  justify-content: space-between;
+`;
 
 const Displaying = styled("span")`
   font-size: ${({ theme }) => theme.font.xs};
@@ -27,34 +73,17 @@ const StyledButton = styled(TransparentButton)`
   margin: 3px 0 3px 5px;
 `;
 
-const ConceptTreeSearchBox = ({
-  showMismatches,
-  onToggleShowMismatches,
-  areTreesAvailable,
-  ...props
-}) => {
-  if (!areTreesAvailable) return null;
+type PropsT = {
+  datasetId: string,
 
-  return (
-    <SearchBox
-      {...props}
-      placeholder={T.translate("conceptTreeList.searchPlaceholder")}
-      textAppend={
-        <div>
-          <Displaying>
-            {showMismatches
-              ? T.translate("conceptTreeList.showingMismatches")
-              : T.translate("conceptTreeList.showingMatchesOnly")}
-          </Displaying>
-          <StyledButton tiny onClick={onToggleShowMismatches}>
-            {showMismatches
-              ? T.translate("conceptTreeList.showMatchesOnly")
-              : T.translate("conceptTreeList.showMismatches")}
-          </StyledButton>
-        </div>
-      }
-    />
-  );
+  search: SearchT,
+  areTreesAvailable: boolean,
+  showMismatches: boolean,
+
+  onSearch: (datasetId: DatasetIdT, value: string) => void,
+  onChange: string => void,
+  onClearQuery: () => void,
+  onToggleShowMismatches: () => void
 };
 
 const mapStateToProps = state => ({
@@ -67,7 +96,6 @@ const mapDispatchToProps = dispatch => ({
   onSearch: (datasetId, query) => {
     if (query.length > 1) dispatch(searchTrees(datasetId, query));
   },
-  onChange: query => dispatch(changeSearchQuery(query)),
   onClearQuery: () => dispatch(clearSearchQuery()),
   onToggleShowMismatches: () => dispatch(toggleShowMismatches())
 });
@@ -75,4 +103,79 @@ const mapDispatchToProps = dispatch => ({
 export default connect(
   mapStateToProps,
   mapDispatchToProps
-)(ConceptTreeSearchBox);
+)(
+  ({
+    datasetId,
+    search,
+    onSearch,
+    onClearQuery,
+    showMismatches,
+    onToggleShowMismatches,
+    areTreesAvailable
+  }: PropsT) => {
+    const [localQuery, setLocalQuery] = React.useState("");
+
+    React.useEffect(() => {
+      setLocalQuery(search.query);
+    }, [search.query]);
+
+    if (!areTreesAvailable) return null;
+
+    return (
+      <Root>
+        <StyledBaseInput
+          placeholder={T.translate("conceptTreeList.searchPlaceholder")}
+          value={localQuery || ""}
+          onChange={value => {
+            if (isEmpty(value)) onClearQuery();
+
+            setLocalQuery(value);
+          }}
+          inputProps={{
+            onKeyPress: e => {
+              return e.key === "Enter" && !isEmpty(e.target.value)
+                ? onSearch(datasetId, e.target.value)
+                : null;
+            }
+          }}
+        />
+        {!isEmpty(localQuery) && (
+          <Right>
+            <StyledIconButton
+              icon="search"
+              aria-hidden="true"
+              onClick={() => onSearch(datasetId, localQuery)}
+            />
+          </Right>
+        )}
+        {search.loading ? (
+          <AnimatedDots />
+        ) : (
+          search.result &&
+          search.resultCount >= 0 && (
+            <Row>
+              <TinyText>
+                {T.translate("search.resultLabel", {
+                  totalResults: search.resultCount,
+                  duration: (search.duration / 1000.0).toFixed(2)
+                })}
+              </TinyText>
+              <div>
+                <Displaying>
+                  {showMismatches
+                    ? T.translate("conceptTreeList.showingMismatches")
+                    : T.translate("conceptTreeList.showingMatchesOnly")}
+                </Displaying>
+                <StyledButton tiny onClick={onToggleShowMismatches}>
+                  {showMismatches
+                    ? T.translate("conceptTreeList.showMatchesOnly")
+                    : T.translate("conceptTreeList.showMismatches")}
+                </StyledButton>
+              </div>
+            </Row>
+          )
+        )}
+      </Root>
+    );
+  }
+);
