+++ frontend/lib/js/concept-trees/actionTypes.js
@@ -14,7 +14,6 @@ export const LOAD_TREE_ERROR = "concept-trees/LOAD_TREE_ERROR";
 export const SEARCH_TREES_START = "concept-trees/SEARCH_TREES_START";
 export const SEARCH_TREES_SUCCESS = "concept-trees/SEARCH_TREES_SUCCESS";
 export const SEARCH_TREES_ERROR = "concept-trees/SEARCH_TREES_ERROR";
-export const CHANGE_SEARCH_QUERY = "concept-trees/CHANGE_SEARCH_QUERY";
 export const CLEAR_SEARCH_QUERY = "concept-trees/CLEAR_SEARCH_QUERY";
 
 export const TOGGLE_ALL_OPEN = "concept-trees/TOGGLE_ALL_OPEN";
